Minor code cleanup / reorg
