+++ Co-Badged Cards/SwiftUI/Co-Badged Cards Example/Models/PrimerCardDataModel.swift
@@ -91,43 +91,3 @@ extension PrimerCardDataModel: PrimerDataServiceModelsDelegate {
         }
     }
 }
-
-// MARK: Errors Model
-
-class PrimerCardDataErrorsModel: PrimerBaseCardDataModel {
-    
-    fileprivate func clearErrors() {
-        self.cardNumber = ""
-        self.expiryDate = ""
-        self.cvvNumber = ""
-        self.cardholderName = ""
-        self.selectedCardNetwork = .unknown
-    }
-}
-
-extension PrimerCardDataErrorsModel: PrimerDataServiceErrorsDelegate {
-    func didReceiveErrors(errors: [Error]) {
-        DispatchQueue.main.async {
-            self.clearErrors()
-            
-            let validationErrors = errors.reversed().compactMap { $0 as? PrimerValidationError }
-            
-            validationErrors.forEach { error in
-                logger.info("[PrimerCardDataModel.didReceiveErrors] => \(error.errorId)")
-                switch error {
-                case .invalidCardnumber(let message, _, _):
-                    self.cardNumber = message
-                case .invalidCardType(let message, _, _):
-                    self.cardNumber = message // Overrides above
-                case .invalidExpiryDate(let message, _, _):
-                    self.expiryDate = message
-                case .invalidCvv(let message, _, _):
-                    self.cvvNumber = message
-                case .invalidCardholderName(let message, _, _):
-                    self.cardholderName = message
-                default: break
-                }
-            }
-        }
-    }
-}
